Silence format warning on macOS
The struct stat member st_nlink on macOS is of the type unsigned short.
Use an explicit cast.

    logrotate.c:1406:37: error: format specifies type 'unsigned long' but the argument has type 'nlink_t' (aka 'unsigned short') [-Werror,-Wformat]
                    log->files[logNum], sb.st_nlink, log->pattern);
                                        ^~~~~~~~~~~

+++ logrotate.c
@@ -1403,7 +1403,7 @@ static int findNeedRotating(const struct logInfo *log, unsigned logNum, int forc
     if (!(log->flags & LOG_FLAG_ALLOWHARDLINK) && sb.st_nlink != 1) {
         message(MESS_DEBUG, "  log %s has multiple (%lu) hard links. Rotation of files"
                 " with multiple hard links is not allowed for %s -- skipping.\n",
-                log->files[logNum], sb.st_nlink, log->pattern);
+                log->files[logNum], (unsigned long)sb.st_nlink, log->pattern);
         return 0;
     }
